Rename scifinder_credential folder to scifinderCredential
