+++ app/packs/src/apps/scifinderCredential/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ScifinderCredential from 'src/apps/scifinder_credential/ScifinderCredential';
+import ScifinderCredential from 'src/apps/scifinderCredential/ScifinderCredential';
 
 document.addEventListener('DOMContentLoaded', () => {
   const domElement = document.getElementById('ScifinderCredential');
